// apps/admin-web/src/api/model/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin API
 * The Admin API description
 * OpenAPI spec version: 1.0
 */
import type { Account } from './account';
import type { Member } from './member';
import type { Session } from './session';
import type { TermAgreement } from './termAgreement';
import type { UserMetadata } from './userMetadata';

export interface User {
  sessions: Session[];
  account?: Account;
  member?: Member;
  termAgreements: TermAgreement[];
  name: string;
  email: string;
  emailVerified: boolean;
  image?: string;
  isAnonymous?: boolean;
  role?: string;
  banned?: boolean;
  banReason?: string;
  banExpires?: string;
  twoFactorEnabled?: boolean;
  twoFactorSecret?: string;
  phoneNumber?: string;
  phoneNumberVerified?: boolean;
  id: string;
  createdAt: string;
  createdBy?: string;
  updatedAt: string;
  updatedBy?: string;
  deletedAt?: string;
  deletedBy?: string;
  metadata?: UserMetadata;
}